# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AnyErrands API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /helpers:
    get:
      operationId: listHelpers
      tags: [helpers]
      summary: List available helpers
      responses:
        "200":
          description: List of helpers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Helper"
    post:
      operationId: createHelper
      tags: [helpers]
      summary: Register as a helper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelperInput"
      responses:
        "201":
          description: Helper profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Helper"
  /helpers/{id}:
    get:
      operationId: getHelper
      tags: [helpers]
      summary: Get helper profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Helper profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Helper"
        "404":
          description: Not found
    patch:
      operationId: updateHelper
      tags: [helpers]
      summary: Update helper profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelperUpdate"
      responses:
        "200":
          description: Updated helper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Helper"
  /helpers/{id}/errands:
    get:
      operationId: getHelperErrands
      tags: [helpers]
      summary: Get errands accepted by a helper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Helper's errands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Errand"
  /helpers/{id}/earnings:
    get:
      operationId: getHelperEarnings
      tags: [helpers]
      summary: Get a helper's earnings and completed paid jobs (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Helper earnings summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelperEarnings"
        "403":
          description: Not the owner of this helper profile
        "404":
          description: Not found
  /helpers/{id}/reviews:
    get:
      operationId: getHelperReviews
      tags: [helpers]
      summary: Get reviews left for a helper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Helper's reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelperReviews"
...
  /welcome:
    get:
      operationId: getWelcomeStatus
      tags: [welcome]
      summary: Whether the current user should see a welcome popup
      responses:
        "200":
          description: Welcome status for the current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WelcomeStatus"
        "401":
          description: Not logged in
  /welcome/customer/seen:
    post:
      operationId: markCustomerWelcomeSeen
      tags: [welcome]
      summary: Dismiss the customer welcome popup for the current user
      responses:
        "200":
          description: Updated welcome status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WelcomeStatus"
        "401":
          description: Not logged in
  /welcome/helper/seen:
    post:
      operationId: markHelperWelcomeSeen
      tags: [welcome]
      summary: Dismiss the helper welcome popup for the current user
      responses:
        "200":
          description: Updated welcome status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WelcomeStatus"
        "401":
          description: Not logged in
components:
...
  schemas:
...
    WelcomeStatus:
      type: object
      required: [showCustomerWelcome, showHelperWelcome, firstName]
      properties:
        showCustomerWelcome:
          type: boolean
        showHelperWelcome:
          type: boolean
        firstName:
          type: [string, 'null']
...
    Errand:
      type: object
      required: [id, title, description, category, status, requesterName, requesterLocation, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [open, accepted, completed]
        requesterName:
          type: string
        requesterLocation:
          type: string
        requesterPhone:
          type: ["string", "null"]
        estimatedDuration:
          type: ["string", "null"]
        budgetAmount:
          type: ["number", "null"]
        tripFrom:
          type: ["string", "null"]
        tripTo:
          type: ["string", "null"]
        tripWhen:
          type: ["string", "null"]
        passengers:
          type: ["integer", "null"]
        returnTrip:
          type: boolean
        helperId:
          type: ["integer", "null"]
        helperName:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [unpaid, paid, refunded]
        paidAmount:
          type: ["number", "null"]
        platformFee:
          type: ["number", "null"]
        paidAt:
          type: ["string", "null"]
        helperPaidAt:
          type: ["string", "null"]
          description: When the held payment was released to the helper.
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
        isRequester:
          type: boolean
          description: True when the authenticated user is the person who posted this errand.
...
    Helper:
      type: object
      required: [id, name, location, bio, skills, errandsCompleted, rating, available, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        errandsCompleted:
          type: integer
        rating:
          type: ["number", "null"]
        available:
          type: boolean
        avatarInitials:
          type: ["string", "null"]
        stripeAccountId:
          type: ["string", "null"]
        stripeOnboardingComplete:
          type: ["boolean", "null"]
        isOwner:
          type: boolean
          description: True when the currently authenticated user owns this helper profile
        createdAt:
          type: string
    HelperEarningJob:
      type: object
      required: [errandId, title, paidAmount, platformFee, earned, completedAt]
      properties:
        errandId:
          type: integer
        title:
          type: string
        paidAmount:
          type: number
        platformFee:
          type: number
        earned:
          type: number
        completedAt:
          type: ["string", "null"]
    HelperEarnings:
      type: object
      required: [totalEarned, totalPaidOut, platformFeesPaid, jobsCount, jobs]
      properties:
        totalEarned:
          type: number
          description: Sum of what the helper actually earned (paid minus platform fee)
        totalPaidOut:
          type: number
          description: Total amount requesters paid across the helper's paid jobs
        platformFeesPaid:
          type: number
          description: Total platform fees deducted across the helper's paid jobs
        jobsCount:
          type: integer
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/HelperEarningJob"
    Review:
      type: object
      required: [id, errandId, helperId, reviewerName, rating, comment, createdAt]
      properties:
        id:
          type: integer
        errandId:
          type: integer
        helperId:
          type: integer
        reviewerName:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    HelperReviews:
      type: object
      required: [averageRating, reviewCount, reviews]
      properties:
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
...
    HelperInput:
      type: object
      required: [name, location, bio, skills]
      properties:
        name:
          type: string
        location:
          type: string
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        available:
          type: boolean
    HelperUpdate:
      type: object
      properties:
        bio:
          type: string
        location:
          type: string
        skills:
          type: array
          items:
            type: string
        available:
          type: boolean
...