# lib/api-zod/src/generated/types/welcomeStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnyErrands API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WelcomeStatus {
  showCustomerWelcome: boolean;
  showHelperWelcome: boolean;
  /** @nullable */
  firstName: string | null;
}